backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
sqlalchemy==2.0.27
aiosqlite==0.20.0
python-dotenv==1.0.0
openai==1.16.1 